Refactor: RP-106 svg -> svgr 리팩토링
